Namespaced classes etc as per issue #19
