+++ tests/test_get_oauth_params.rb
@@ -1,6 +1,6 @@
 require 'minitest/autorun'
 require 'minitest/mock'
-require_relative '../lib/oauth'
+require_relative '../lib/mastercard/oauth'
 
 class TestGetOAuthParams < Minitest::Test
 
@@ -8,14 +8,14 @@ class TestGetOAuthParams < Minitest::Test
   MY_PAYLOAD = '{ my: "payload" }'.freeze
 
   def test_creates_map_with_ordered_oauth_parameters
-    oauth_params = OAuth.get_oauth_params(CONSUMER_KEY)
+    oauth_params = Mastercard::OAuth.get_oauth_params(CONSUMER_KEY)
     map_keys = oauth_params.keys
     params = %w[oauth_body_hash oauth_consumer_key oauth_nonce oauth_signature_method oauth_timestamp oauth_version]
     assert_equal(map_keys, params)
   end
 
   def test_creates_map_with_ordered_oauth_parameters_with_payload
-    oauth_params = OAuth.get_oauth_params(CONSUMER_KEY, MY_PAYLOAD)
+    oauth_params = Mastercard::OAuth.get_oauth_params(CONSUMER_KEY, MY_PAYLOAD)
     map_keys = oauth_params.keys
     params = %w[oauth_body_hash oauth_consumer_key oauth_nonce oauth_signature_method oauth_timestamp oauth_version]
     assert_equal(map_keys, params)
